developing

+++ TIN/src/main/java/com/springapp/mvc/controllers/TinController.java
@@ -48,31 +48,43 @@ public class TinController
         ModelAndView modelAndView = new ModelAndView();
         if(singInForm.getUsername().equals(App.ADMIN_USERNAME) && singInForm.getPassword().equals(App.ADMIN_PASSWORD)){
             singInForm.setWho(SingInForm.Who.Admin);
+            session.setWho(UserSession.Who.Admin);
             session.setId(0);
         } else {
             List<User> users = database.getUserByMail(singInForm.getUsername());
             if(!users.isEmpty()){
                 user = database.getUserByMail(singInForm.getUsername()).get(0);
                 singInForm.setWho(SingInForm.Who.Promoter);
+                session.setWho(UserSession.Who.Promoter);
                 session.setId(user.getId());
             } else {
                 singInForm.setWho(SingInForm.Who.None);
             }
         }
-        switch (singInForm.getWho())
+        switch (session.getWho())
         {
             case Promoter:
                 return "redirect:" + App.BASKETS;
             case Admin:
                 return "redirect:.." + App.ADMIN_CONSTROLLER_URL + App.ADMIN;
             default:
-                return "redirect:.." +App.TIN_CONTROLLER_URL;
+                return "redirect:.." + App.TIN_CONTROLLER_URL;
         }
+    }
+
+    @RequestMapping(value = App.LOG_OUT, method = RequestMethod.GET)
+    public String logOut()
+    {
+        session.setWho(UserSession.Who.None);
+        return  "redirect:" + App.TIN_CONTROLLER_URL;
     }
 
     @RequestMapping(value = App.BASKETS, method = RequestMethod.GET)
     public ModelAndView basketsPage()
     {
+        if(!hasPermission()){
+            return new ModelAndView(App.TIN_CONTROLLER_URL);
+        }
         ModelAndView modelAndView = new ModelAndView();
         modelAndView.setViewName(App.BASKETS);
         modelAndView.addObject("baskets", database.getBaskets(session.getId()));
@@ -82,6 +94,9 @@ public class TinController
     @RequestMapping(value = App.BASKET)
     public ModelAndView homeTin(HttpServletRequest request)
     {
+        if(!hasPermission()){
+            return new ModelAndView(App.TIN_CONTROLLER_URL);
+        }
         String idBasket = request.getParameter(App.ID_EXIST_BASKET_PARAMETER);
         int id = Integer.valueOf(idBasket);
 
@@ -93,9 +108,11 @@ public class TinController
     @RequestMapping(value = App.BASKETS, method = RequestMethod.POST)
     public String addNewBasket(HttpServletRequest request)
     {
+        if(!hasPermission()){
+            return "redirect:" + App.TIN_CONTROLLER_URL;
+        }
         String basketName = request.getParameter(App.NAME_NEW_BASKET_PARAMETER);
         User user = database.getUser(session.getId());
-//        User user = database.getUser(1);
         Basket basket = new Basket(basketName, user);
         database.saveBasket(basket);
         return "redirect:" + App.BASKETS;
@@ -104,6 +121,9 @@ public class TinController
     @RequestMapping(value = App.DELETE_BASKET, method = RequestMethod.POST)
     public String deleteBasket(HttpServletRequest request)
     {
+        if(!hasPermission()){
+            return "redirect:" + App.TIN_CONTROLLER_URL;
+        }
         String idBasket = request.getParameter(App.ID_EXIST_BASKET_PARAMETER);
         int id = Integer.valueOf(idBasket);
         database.deleteBasket(session.getId(), id);
@@ -113,6 +133,9 @@ public class TinController
     @RequestMapping(value = App.EDIT_BASKET, method = RequestMethod.POST)
     public String editBasket(HttpServletRequest request)
     {
+        if(!hasPermission()){
+            return "redirect:" + App.TIN_CONTROLLER_URL;
+        }
         String idBasket = request.getParameter(App.ID_EXIST_BASKET_PARAMETER);
         String newBasketName = request.getParameter(App.NAME_NEW_BASKET_PARAMETER);
         int id = Integer.valueOf(idBasket);
@@ -125,6 +148,9 @@ public class TinController
     @RequestMapping(value = App.GET_RECORDS_JSON, method = RequestMethod.POST)
     public @ResponseBody List<Record> getRecords(@RequestParam(value = "" + App.ID_EXIST_BASKET_PARAMETER) Integer idBasket)
     {
+        if(!hasPermission()){
+            return new LinkedList<Record>();
+        }
         List<Record> tmp = database.getRecordsByBasket(session.getId(), idBasket);
         return tmp;
     }
@@ -132,7 +158,10 @@ public class TinController
     @RequestMapping(value = App.SAVE_RECORDS, method = RequestMethod.POST)
     public ModelAndView saveRecords(@RequestBody LinkedList<Record> records)
     {
-        //TODO
+        //TODO NOT WORKS
+        if(!hasPermission()){
+            return new ModelAndView(App.TIN_CONTROLLER_URL);
+        }
         List<Record> list = new ArrayList<>();
         database.saveRecordsByBasket(session.getId(), 1, list);
         ModelAndView modelAndView = new ModelAndView("basket");
@@ -142,6 +171,9 @@ public class TinController
     @RequestMapping(value = App.ADD_RECORD, method = RequestMethod.POST)
     public @ResponseBody Integer addRecord(@RequestParam(value = "" + App.ID_EXIST_BASKET_PARAMETER) Integer idBasket)
     {
+        if(!hasPermission()){
+            return new Integer(-1);
+        }
         Record record = new Record();
         return (int)(long)database.addNewRecord(session.getId(), idBasket, record).getId();
     }
@@ -150,6 +182,9 @@ public class TinController
     public @ResponseBody Record getRecords(@RequestParam(value = "" + App.ID_EXIST_BASKET_PARAMETER) Integer idBasket,
                                            @RequestParam(value = "" + App.ID_EXIST_RECORD_PARAMETER) Integer idRecord)
     {
+        if(!hasPermission()){
+            return new Record();
+        }
         Record record = database.getRecord(session.getId(), idBasket, idRecord);
         return record;
     }
@@ -157,6 +192,9 @@ public class TinController
     @RequestMapping(value = App.DELETE_RECORD, method = RequestMethod.POST)
     public @ResponseBody Boolean deleteRecord(HttpServletRequest request)
     {
+        if(!hasPermission()){
+            return new Boolean(false);
+        }
         int idBasket = Integer.valueOf(request.getParameter(App.ID_EXIST_BASKET_PARAMETER));
         int idRecord = Integer.valueOf(request.getParameter(App.ID_EXIST_RECORD_PARAMETER));
         if(database.deleteRecord(session.getId(), idBasket, idRecord))
@@ -167,6 +205,9 @@ public class TinController
     @RequestMapping(value = App.SAVE_RECORD, method = RequestMethod.POST)
     public @ResponseBody Boolean saveRecord(@RequestParam("idBasket") Integer idBasket,@RequestBody Record record)
     {
+        if(!hasPermission()){
+            return new Boolean(false);
+        }
         if(database.saveRecord(session.getId(), idBasket, record) != null)
             return new Boolean(true);
         return new Boolean(false);
@@ -175,6 +216,9 @@ public class TinController
     @RequestMapping(value = App.GENERATE_XML_INZ)
     public @ResponseBody String generateXMLInz(@RequestParam("idBasket") Integer idBasket)
     {
+        if(!hasPermission()){
+            return "You don't have permission";
+        }
         List<Record> records = database.getRecordsByBasket(idBasket);
         String ret = "Records";
         for(Record rec: records){
@@ -187,6 +231,9 @@ public class TinController
     @RequestMapping(value = App.GENERATE_XML_MGR)
     public @ResponseBody String generateXMLMgr(@RequestParam("idBasket") Integer idBasket)
     {
+        if(!hasPermission()){
+            return "You don't have permission";
+        }
         List<Record> records = database.getRecordsByBasket(idBasket);
         String ret = "Records";
         for(Record rec: records){
@@ -194,5 +241,12 @@ public class TinController
             //TODO Jan tu zaimplementuj "ret += funckcja(rec)" dla prac magisterskich
         }
         return ret;
+    }
+
+    boolean hasPermission()
+    {
+        if(session.getWho().equals(UserSession.Who.None))
+            return false;
+        return true;
     }
 }
